fix(e2e): retry the pre-fork etcd probe, and only demand the clusters in use
Three review findings.

globalSetup probed each endpoint ONCE. The 5s budget bought no
cold-start tolerance at all, because a container still opening its
listener REFUSES the connection in milliseconds rather than hanging — so
one blip on any of the four services fails the whole merge-blocking e2e
leg before a single test runs, which is a worse failure than the silent
skip it was added to prevent. It now escalates over five attempts, the
way ping() already does on CI and for the same reason.

It also demanded every CONFIGURED endpoint, while forkBudget is capped
by the core count as well — so on a smaller runner a four-entry list
drives two forks and the run would have failed over two clusters it was
never going to touch. It now probes the leading slice that
`poolId % list.length` will actually reach.

The ci.yml comment claimed adding a service raises concurrency. It does
not: ubuntu-latest has exactly four cores, so a fifth cluster would add
no forks and only a fifth thing required to be up. Corrected to say a
bigger runner has to come first.

forkBudget and the endpoint-list parsing move to harness/forks.ts, which
is now the single place either is derived — etcd.ts asks it rather than
re-splitting the variable. That shrinks the endpoint guard's allowlist
instead of growing it: two files may resolve an endpoint, forks.ts for
the list and etcd.ts for the fallback and the fork mapping. (The guard
caught forks.ts itself on the first pass, which is what prompted routing
it this way rather than adding an exemption.)

Also dropped file-resource-source-e2e's own `if (process.env.CI) throw`.
ping() now throws on CI for every case file, so that branch was
unreachable — and the rule it encoded is the one that moved into the
client.
