+++ tests/e2e/vitest.config.ts
@@ -1,37 +1,6 @@
 import { defineConfig } from "vitest/config";
 
-import { availableParallelism } from "node:os";
-
-import { RANGE_SLOTS } from "./src/harness/ports.js";
-
-/**
- * One fork per etcd cluster the run was given.
- *
- * Floor 2 — the value this suite ran at before the clusters were split,
- * and what a developer with a single local etcd gets. Note the floor
- * means ONE configured endpoint still yields two forks sharing it; that
- * is the pre-existing arrangement, not a regression, but it is also why
- * "fewer endpoints => fewer forks" only holds above 2.
- *
- * Ceiling RANGE_SLOTS — ports.ts hands fork `poolId % RANGE_SLOTS` its
- * own port range, so beyond that two forks collide on a range and the
- * AddrInUse flake that module exists to prevent comes back.
- *
- * Also ceilinged by the core count. Each fork drives a real `aisix`
- * process, and several cases assert wall-clock bounds derived on a
- * loaded runner (audio-timeout, ttft-first-frame,
- * per-attempt-telemetry) — handing out more forks than cores would eat
- * their margin whatever the endpoint list says.
- */
-function forkBudget(): number {
-  const ceiling = Math.min(RANGE_SLOTS, Math.max(2, availableParallelism()));
-  const override = Number(process.env.AISIX_E2E_MAX_FORKS);
-  if (Number.isInteger(override) && override >= 1) return Math.min(override, ceiling);
-  const endpoints = (process.env.AISIX_E2E_ETCD_ENDPOINTS ?? "")
-    .split(",")
-    .filter((s) => s.trim() !== "").length;
-  return Math.min(Math.max(2, endpoints), ceiling);
-}
+import { forkBudget } from "./src/harness/forks.js";
 
 export default defineConfig({
   test: {
